Refactor CraftingRecipeLogic to use RecipeTransferSyncHandler
